ISPN-8078 Refactor helpers out of NonTxDistributionInterceptor
* Also move BaseDistributionInterceptor$ArrayIterator -> commons/ArrayCollector
