+++ core/src/main/java/org/infinispan/interceptors/distribution/NonTxDistributionInterceptor.java
@@ -1,12 +1,9 @@
 package org.infinispan.interceptors.distribution;
 
 import java.util.ArrayList;
-import java.util.Arrays;
-import java.util.Collection;
 import java.util.Collections;
 import java.util.HashMap;
 import java.util.HashSet;
-import java.util.Iterator;
 import java.util.List;
 import java.util.Map;
 import java.util.Map.Entry;
@@ -14,7 +11,6 @@ import java.util.Set;
 import java.util.concurrent.CompletableFuture;
 import java.util.function.BiConsumer;
 
-import org.infinispan.commands.VisitableCommand;
 import org.infinispan.commands.functional.ReadWriteKeyCommand;
 import org.infinispan.commands.functional.ReadWriteKeyValueCommand;
 import org.infinispan.commands.functional.ReadWriteManyCommand;
@@ -36,8 +32,6 @@ import org.infinispan.context.InvocationContext;
 import org.infinispan.context.impl.FlagBitSets;
 import org.infinispan.distribution.LocalizedCacheTopology;
 import org.infinispan.distribution.ch.ConsistentHash;
-import org.infinispan.distribution.util.ReadOnlySegmentAwareCollection;
-import org.infinispan.distribution.util.ReadOnlySegmentAwareMap;
 import org.infinispan.interceptors.InvocationFinallyAction;
 import org.infinispan.interceptors.InvocationSuccessFunction;
 import org.infinispan.remoting.responses.SuccessfulResponse;
@@ -69,35 +63,11 @@ public class NonTxDistributionInterceptor extends BaseDistributionInterceptor {
    private static Log log = LogFactory.getLog(NonTxDistributionInterceptor.class);
    private static final boolean trace = log.isTraceEnabled();
 
-   private final PutMapHelper putMapHelper = new PutMapHelper();
-   private final ReadWriteManyHelper readWriteManyHelper = new ReadWriteManyHelper();
-   private final ReadWriteManyEntriesHelper readWriteManyEntriesHelper = new ReadWriteManyEntriesHelper();
-   private final WriteOnlyManyEntriesHelper writeOnlyManyEntriesHelper = new WriteOnlyManyEntriesHelper();
-   private final WriteOnlyManyHelper writeOnlyManyHelper = new WriteOnlyManyHelper();
-
-   private static BiConsumer<MergingCompletableFuture<Object>, Object> moveListItemsToFuture(int myOffset) {
-      return (f, rv) -> {
-         Collection<?> items;
-         if (rv == null && f.results == null) {
-            return;
-         } else if (rv instanceof Map) {
-            items = ((Map) rv).entrySet();
-         } else if (rv instanceof Collection) {
-            items = (Collection<?>) rv;
-         } else {
-            f.completeExceptionally(new IllegalArgumentException("Unexpected result value " + rv));
-            return;
-         }
-         if (trace) {
-            log.tracef("Copying %d items %s to results (%s), starting offset %d", items.size(), items,
-                  Arrays.toString(f.results), myOffset);
-         }
-         Iterator<?> it = items.iterator();
-         for (int i = 0; it.hasNext(); ++i) {
-            f.results[myOffset + i] = it.next();
-         }
-      };
-   }
+   private final PutMapHelper putMapHelper = new PutMapHelper(this::createRemoteCallback);
+   private final ReadWriteManyHelper readWriteManyHelper = new ReadWriteManyHelper(this::createRemoteCallback);
+   private final ReadWriteManyEntriesHelper readWriteManyEntriesHelper = new ReadWriteManyEntriesHelper(this::createRemoteCallback);
+   private final WriteOnlyManyEntriesHelper writeOnlyManyEntriesHelper = new WriteOnlyManyEntriesHelper(this::createRemoteCallback);
+   private final WriteOnlyManyHelper writeOnlyManyHelper = new WriteOnlyManyHelper(this::createRemoteCallback);
 
    private Map<Address, Set<Integer>> primaryOwnersOfSegments(ConsistentHash ch) {
       Map<Address, Set<Integer>> map = new HashMap<>(ch.getMembers().size());
@@ -205,7 +175,7 @@ public class NonTxDistributionInterceptor extends BaseDistributionInterceptor {
       ConsistentHash ch = cacheTopology.getWriteConsistentHash();
       if (ctx.isOriginLocal()) {
          Map<Address, Set<Integer>> segmentMap = primaryOwnersOfSegments(ch);
-         CountDownCompletableFuture allFuture = new CountDownCompletableFuture(ctx, segmentMap.size());
+         CountDownCompletableFuture allFuture = new CountDownCompletableFuture(segmentMap.size());
 
          // Go through all members, for this node invokeNext (if this node is an owner of some keys),
          // for the others (that own some keys) issue a remote call.
@@ -268,7 +238,7 @@ public class NonTxDistributionInterceptor extends BaseDistributionInterceptor {
       }
 
       if (helper.shouldRegisterRemoteCallback(command)) {
-         return invokeNextThenApply(ctx, command, helper);
+         return invokeNextThenApply(ctx, command, helper.remoteCallback);
       } else {
          return invokeNext(ctx, command);
       }
@@ -307,7 +277,7 @@ public class NonTxDistributionInterceptor extends BaseDistributionInterceptor {
             results = new Object[helper.getItems(command).size()];
          }
          MergingCompletableFuture<Object> allFuture
-               = new MergingCompletableFuture<>(ctx, segmentMap.size(), results, helper::transformResult);
+               = new MergingCompletableFuture<>(segmentMap.size(), results, helper::transformResult);
          MutableInt offset = new MutableInt();
 
          // Go through all members, for this node invokeNext (if this node is an owner of some keys),
@@ -352,7 +322,7 @@ public class NonTxDistributionInterceptor extends BaseDistributionInterceptor {
 
       C localCommand = helper.copyForLocal(command, myItems);
       InvocationFinallyAction handler =
-            createLocalInvocationHandler(ch, allFuture, segments, helper, moveListItemsToFuture(myOffset));
+            createLocalInvocationHandler(ch, allFuture, segments, helper, MergingCompletableFuture.moveListItemsToFuture(myOffset));
       if (retrievals == null) {
          invokeNextAndFinally(ctx, localCommand, handler);
       } else {
@@ -391,7 +361,7 @@ public class NonTxDistributionInterceptor extends BaseDistributionInterceptor {
                      return;
                   }
                   Object responseValue = response.getResponseValue();
-                  moveListItemsToFuture(myOffset).accept(allFuture, responseValue);
+                  MergingCompletableFuture.moveListItemsToFuture(responseValue, allFuture, myOffset);
                   allFuture.countDown();
                }
             });
@@ -414,7 +384,7 @@ public class NonTxDistributionInterceptor extends BaseDistributionInterceptor {
       }
       Object result = asyncInvokeNext(ctx, command, delay);
       if (helper.shouldRegisterRemoteCallback(command)) {
-         return makeStage(result).thenApply(ctx, command, helper);
+         return makeStage(result).thenApply(ctx, command, helper.remoteCallback);
       } else {
          return result;
       }
@@ -492,333 +462,28 @@ public class NonTxDistributionInterceptor extends BaseDistributionInterceptor {
       public int value;
    }
 
-   private abstract class WriteManyCommandHelper<C extends WriteCommand, Container, Item>
-         implements InvocationSuccessFunction {
-      public abstract C copyForLocal(C cmd, Container container);
-
-      public abstract C copyForPrimary(C cmd, ConsistentHash ch, Set<Integer> segments);
-
-      public abstract C copyForBackup(C cmd, ConsistentHash ch, Set<Integer> segments);
-
-      public abstract Collection<Item> getItems(C cmd);
-
-      public abstract Object item2key(Item item);
-
-      public abstract Container newContainer();
-
-      public abstract void accumulate(Container container, Item item);
-
-      public abstract int containerSize(Container container);
-
-      public abstract boolean shouldRegisterRemoteCallback(C cmd);
-
-      public abstract Object transformResult(Object[] results);
-
-      @Override
-      public Object apply(InvocationContext rCtx, VisitableCommand rCommand, Object rv) throws Throwable {
-         C original = (C) rCommand;
-         ConsistentHash ch = checkTopologyId(original).getWriteConsistentHash();
-         // We have already checked that the command topology is actual, so we can assume that we really are primary owner
-         Map<Address, Set<Integer>> backups = backupOwnersOfSegments(ch, ch.getPrimarySegmentsForOwner(rpcManager.getAddress()));
-         if (backups.isEmpty()) {
-            return rv;
+   private <C extends WriteCommand> Object writeManyRemoteCallback(WriteManyCommandHelper<C, ?, ?> helper, InvocationContext ctx, C command, Object rv) {
+      ConsistentHash ch = checkTopologyId(command).getWriteConsistentHash();
+      // We have already checked that the command topology is actual, so we can assume that we really are primary owner
+      Map<Address, Set<Integer>> backups = backupOwnersOfSegments(ch, ch.getPrimarySegmentsForOwner(rpcManager.getAddress()));
+      if (backups.isEmpty()) {
+         return rv;
+      }
+      boolean isSync = isSynchronous(command);
+      CompletableFuture[] futures = isSync ? new CompletableFuture[backups.size()] : null;
+      int future = 0;
+      for (Entry<Address, Set<Integer>> backup : backups.entrySet()) {
+         C copy = helper.copyForBackup(command, ch, backup.getValue());
+         if (isSync) {
+            futures[future++] = rpcManager.invokeRemotelyAsync(Collections.singleton(backup.getKey()), copy, defaultSyncOptions);
+         } else {
+            rpcManager.invokeRemotelyAsync(Collections.singleton(backup.getKey()), copy, defaultAsyncOptions);
          }
-         boolean isSync = isSynchronous(original);
-         CompletableFuture[] futures = isSync ? new CompletableFuture[backups.size()] : null;
-         int future = 0;
-         for (Entry<Address, Set<Integer>> backup : backups.entrySet()) {
-            C copy = copyForBackup(original, ch, backup.getValue());
-            if (isSync) {
-               futures[future++] = rpcManager.invokeRemotelyAsync(Collections.singleton(backup.getKey()), copy, defaultSyncOptions);
-            } else {
-               rpcManager.invokeRemotelyAsync(Collections.singleton(backup.getKey()), copy, defaultAsyncOptions);
-            }
-         }
-         return isSync ? asyncValue(CompletableFuture.allOf(futures).thenApply(nil -> rv)) : rv;
       }
+      return isSync ? asyncValue(CompletableFuture.allOf(futures).thenApply(nil -> rv)) : rv;
    }
 
-   private class PutMapHelper extends WriteManyCommandHelper<PutMapCommand, Map<Object, Object>, Entry<Object, Object>> {
-      @Override
-      public PutMapCommand copyForLocal(PutMapCommand cmd, Map<Object, Object> container) {
-         return new PutMapCommand(cmd).withMap(container);
-      }
-
-      @Override
-      public PutMapCommand copyForPrimary(PutMapCommand cmd, ConsistentHash ch, Set<Integer> segments) {
-         return new PutMapCommand(cmd).withMap(new ReadOnlySegmentAwareMap<>(cmd.getMap(), ch, segments));
-      }
-
-      @Override
-      public PutMapCommand copyForBackup(PutMapCommand cmd, ConsistentHash ch, Set<Integer> segments) {
-         PutMapCommand copy = new PutMapCommand(cmd).withMap(new ReadOnlySegmentAwareMap(cmd.getMap(), ch, segments));
-         copy.setForwarded(true);
-         return copy;
-      }
-
-      @Override
-      public Collection<Entry<Object, Object>> getItems(PutMapCommand cmd) {
-         return cmd.getMap().entrySet();
-      }
-
-      @Override
-      public Object item2key(Entry<Object, Object> entry) {
-         return entry.getKey();
-      }
-
-      @Override
-      public Map<Object, Object> newContainer() {
-         return new HashMap<>();
-      }
-
-      @Override
-      public void accumulate(Map<Object, Object> map, Entry<Object, Object> entry) {
-         map.put(entry.getKey(), entry.getValue());
-      }
-
-      @Override
-      public int containerSize(Map<Object, Object> map) {
-         return map.size();
-      }
-
-      @Override
-      public boolean shouldRegisterRemoteCallback(PutMapCommand cmd) {
-         return !cmd.isForwarded();
-      }
-
-      @Override
-      public Object transformResult(Object[] results) {
-         if (results == null) return null;
-         Map<Object, Object> result = new HashMap<>();
-         for (Object r : results) {
-            Map.Entry<Object, Object> entry = (Entry<Object, Object>) r;
-            result.put(entry.getKey(), entry.getValue());
-         }
-         return result;
-      }
-   }
-
-   private class ReadWriteManyEntriesHelper extends WriteManyCommandHelper<ReadWriteManyEntriesCommand, Map<Object, Object>, Entry<Object, Object>> {
-      @Override
-      public ReadWriteManyEntriesCommand copyForLocal(ReadWriteManyEntriesCommand cmd, Map<Object, Object> entries) {
-         return new ReadWriteManyEntriesCommand(cmd).withEntries(entries);
-      }
-
-      @Override
-      public ReadWriteManyEntriesCommand copyForPrimary(ReadWriteManyEntriesCommand cmd, ConsistentHash ch, Set<Integer> segments) {
-         return new ReadWriteManyEntriesCommand(cmd)
-               .withEntries(new ReadOnlySegmentAwareMap<>(cmd.getEntries(), ch, segments));
-      }
-
-      @Override
-      public ReadWriteManyEntriesCommand copyForBackup(ReadWriteManyEntriesCommand cmd, ConsistentHash ch, Set<Integer> segments) {
-         ReadWriteManyEntriesCommand copy = new ReadWriteManyEntriesCommand(cmd)
-               .withEntries(new ReadOnlySegmentAwareMap(cmd.getEntries(), ch, segments));
-         copy.setForwarded(true);
-         return copy;
-      }
-
-      @Override
-      public Collection<Entry<Object, Object>> getItems(ReadWriteManyEntriesCommand cmd) {
-         return cmd.getEntries().entrySet();
-      }
-
-      @Override
-      public Object item2key(Entry<Object, Object> entry) {
-         return entry.getKey();
-      }
-
-      @Override
-      public Map<Object, Object> newContainer() {
-         return new HashMap<>();
-      }
-
-      @Override
-      public void accumulate(Map<Object, Object> map, Entry<Object, Object> entry) {
-         map.put(entry.getKey(), entry.getValue());
-      }
-
-      @Override
-      public int containerSize(Map<Object, Object> map) {
-         return map.size();
-      }
-
-      @Override
-      public boolean shouldRegisterRemoteCallback(ReadWriteManyEntriesCommand cmd) {
-         return !cmd.isForwarded();
-      }
-
-      @Override
-      public Object transformResult(Object[] results) {
-         return results == null ? null : Arrays.asList(results);
-      }
-   }
-
-   private class ReadWriteManyHelper extends WriteManyCommandHelper<ReadWriteManyCommand, Collection<Object>, Object> {
-      @Override
-      public ReadWriteManyCommand copyForLocal(ReadWriteManyCommand cmd, Collection<Object> keys) {
-         return new ReadWriteManyCommand(cmd).withKeys(keys);
-      }
-
-      @Override
-      public ReadWriteManyCommand copyForPrimary(ReadWriteManyCommand cmd, ConsistentHash ch, Set<Integer> segments) {
-         return new ReadWriteManyCommand(cmd).withKeys(new ReadOnlySegmentAwareCollection(cmd.getAffectedKeys(), ch, segments));
-      }
-
-      @Override
-      public ReadWriteManyCommand copyForBackup(ReadWriteManyCommand cmd, ConsistentHash ch, Set<Integer> segments) {
-         ReadWriteManyCommand copy = new ReadWriteManyCommand(cmd).withKeys(
-               new ReadOnlySegmentAwareCollection(cmd.getAffectedKeys(), ch, segments));
-         copy.setForwarded(true);
-         return copy;
-      }
-
-      @Override
-      public Collection<Object> getItems(ReadWriteManyCommand cmd) {
-         return cmd.getAffectedKeys();
-      }
-
-      @Override
-      public Object item2key(Object key) {
-         return key;
-      }
-
-      @Override
-      public Collection<Object> newContainer() {
-         return new ArrayList<>();
-      }
-
-      @Override
-      public void accumulate(Collection<Object> list, Object key) {
-         list.add(key);
-      }
-
-      @Override
-      public int containerSize(Collection<Object> list) {
-         return list.size();
-      }
-
-      @Override
-      public boolean shouldRegisterRemoteCallback(ReadWriteManyCommand cmd) {
-         return !cmd.isForwarded();
-      }
-
-      @Override
-      public Object transformResult(Object[] results) {
-         return results == null ? null : Arrays.asList(results);
-      }
-   }
-
-   private class WriteOnlyManyEntriesHelper extends WriteManyCommandHelper<WriteOnlyManyEntriesCommand, Map<Object, Object>, Entry<Object, Object>> {
-
-      @Override
-      public WriteOnlyManyEntriesCommand copyForLocal(WriteOnlyManyEntriesCommand cmd, Map<Object, Object> entries) {
-         return new WriteOnlyManyEntriesCommand(cmd).withEntries(entries);
-      }
-
-      @Override
-      public WriteOnlyManyEntriesCommand copyForPrimary(WriteOnlyManyEntriesCommand cmd, ConsistentHash ch, Set<Integer> segments) {
-         return new WriteOnlyManyEntriesCommand(cmd)
-               .withEntries(new ReadOnlySegmentAwareMap<>(cmd.getEntries(), ch, segments));
-      }
-
-      @Override
-      public WriteOnlyManyEntriesCommand copyForBackup(WriteOnlyManyEntriesCommand cmd, ConsistentHash ch, Set<Integer> segments) {
-         WriteOnlyManyEntriesCommand copy = new WriteOnlyManyEntriesCommand(cmd)
-               .withEntries(new ReadOnlySegmentAwareMap(cmd.getEntries(), ch, segments));
-         copy.setForwarded(true);
-         return copy;
-      }
-
-      @Override
-      public Collection<Entry<Object, Object>> getItems(WriteOnlyManyEntriesCommand cmd) {
-         return cmd.getEntries().entrySet();
-      }
-
-      @Override
-      public Object item2key(Entry<Object, Object> entry) {
-         return entry.getKey();
-      }
-
-      @Override
-      public Map<Object, Object> newContainer() {
-         return new HashMap<>();
-      }
-
-      @Override
-      public void accumulate(Map<Object, Object> map, Entry<Object, Object> entry) {
-         map.put(entry.getKey(), entry.getValue());
-      }
-
-      @Override
-      public int containerSize(Map<Object, Object> map) {
-         return map.size();
-      }
-
-      @Override
-      public boolean shouldRegisterRemoteCallback(WriteOnlyManyEntriesCommand cmd) {
-         return !cmd.isForwarded();
-      }
-
-      @Override
-      public Object transformResult(Object[] results) {
-         return results == null ? null : Arrays.asList(results);
-      }
-   }
-
-   private class WriteOnlyManyHelper extends WriteManyCommandHelper<WriteOnlyManyCommand, Collection<Object>, Object> {
-      @Override
-      public WriteOnlyManyCommand copyForLocal(WriteOnlyManyCommand cmd, Collection<Object> keys) {
-         return new WriteOnlyManyCommand(cmd).withKeys(keys);
-      }
-
-      @Override
-      public WriteOnlyManyCommand copyForPrimary(WriteOnlyManyCommand cmd, ConsistentHash ch, Set<Integer> segments) {
-         return new WriteOnlyManyCommand(cmd)
-               .withKeys(new ReadOnlySegmentAwareCollection(cmd.getAffectedKeys(), ch, segments));
-      }
-
-      @Override
-      public WriteOnlyManyCommand copyForBackup(WriteOnlyManyCommand cmd, ConsistentHash ch, Set<Integer> segments) {
-         WriteOnlyManyCommand copy = new WriteOnlyManyCommand(cmd)
-               .withKeys(new ReadOnlySegmentAwareCollection(cmd.getAffectedKeys(), ch, segments));
-         copy.setForwarded(true);
-         return copy;
-      }
-
-      @Override
-      public Collection<Object> getItems(WriteOnlyManyCommand cmd) {
-         return cmd.getAffectedKeys();
-      }
-
-      @Override
-      public Object item2key(Object key) {
-         return key;
-      }
-
-      @Override
-      public Collection<Object> newContainer() {
-         return new ArrayList<>();
-      }
-
-      @Override
-      public void accumulate(Collection<Object> list, Object key) {
-         list.add(key);
-      }
-
-      @Override
-      public int containerSize(Collection<Object> list) {
-         return list.size();
-      }
-
-      @Override
-      public boolean shouldRegisterRemoteCallback(WriteOnlyManyCommand cmd) {
-         return !cmd.isForwarded();
-      }
-
-      @Override
-      public Object transformResult(Object[] results) {
-         return results == null ? null : Arrays.asList(results);
-      }
+   private <C extends WriteCommand> InvocationSuccessFunction createRemoteCallback(WriteManyCommandHelper<C, ?, ?> helper) {
+      return (ctx, command, rv) -> writeManyRemoteCallback(helper, ctx, (C) command, rv);
    }
 }
